refactor: remove common code and update functions
