+++ cognite_toolkit/cognite_modules/examples/cdf_functions_dummy/functions/fn_test2/requirements.txt
@@ -1,1 +1,1 @@
-cognite-sdk>=7.13.8
+cognite-sdk>=7.37.0
