+++ tests/tests_unit/test_cdf_tk/run_data/cognite_modules/examples/run_local_functions/functions/fn_test2/handler.py
@@ -2,15 +2,12 @@ from cognite.client import CogniteClient
 
 # You can import from common.tool and get a CDFClientTool instance
 # that can be used to run the function locally and verify capabilities.
-from common.tool import CDFClientTool
 
 
 def handle(data: dict, client: CogniteClient, secrets: dict, function_call_info: dict) -> dict:
-    tool = CDFClientTool(client=client)
     # This will fail unless the function has the specified capabilities.
     print("Print statements will be shown in the logs.")
     print("Running with the following configuration:\n")
-    print(tool)
     return {
         "data": data,
         "secrets": mask_secrets(secrets),
